feat - reorganize qr-dqn
